requirements.txt
appnope==0.1.3
APScheduler==3.6.3
asgiref==3.5.1
asttokens==2.0.5
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
constantly==15.1.0
cryptography==37.0.1
daphne==3.0.2
decorator==5.1.1
Django==4.0.4
django-cors-headers==3.11.0
django-extensions==3.1.5
djangorestframework==3.13.1
executing==0.8.3
factory-boy==3.2.1
Faker==13.11.1
freezegun==1.2.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.3.0
jedi==0.18.1
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.2.1.post0
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
wcwidth==0.2.5
zope.interface==5.4.0
